feat: widen contacts.telegram_id from INT to BIGINT
Telegram user IDs can exceed 2^31 (~2.1 billion), which overflows a
signed 32-bit INT column. Migration 0019 already fixed users.telegram_id;
this applies the same fix to contacts.telegram_id.

Changes:
- Update Contact model to use BigInteger for telegram_id
- Add idempotent Alembic migration 0021
- Add test verifying telegram_id > 2^31 (6_887_867_394) works

+++ backend/app/models/contact.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, JSON, String
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, JSON, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -13,7 +13,7 @@ class Contact(Base):
     project_id: Mapped[int] = mapped_column(ForeignKey("projects.id"), index=True)
     owner_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
     source_id: Mapped[int | None] = mapped_column(ForeignKey("sources.id"), index=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True, index=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True)
     username: Mapped[str | None] = mapped_column(String(255))
     first_name: Mapped[str] = mapped_column(String(255))
     last_name: Mapped[str | None] = mapped_column(String(255))
